+++ common/src/main/java/net/mehvahdjukaar/supplementaries/integration/shulkerboxtooltip/LunchBasketPreviewProvider.java
@@ -5,28 +5,29 @@ import com.misterpemodder.shulkerboxtooltip.api.provider.BlockEntityPreviewProvi
 import com.misterpemodder.shulkerboxtooltip.api.renderer.PreviewRenderer;
 import net.fabricmc.api.EnvType;
 import net.fabricmc.api.Environment;
+import net.mehvahdjukaar.supplementaries.common.block.tiles.LunchBoxBlockTile;
 import net.mehvahdjukaar.supplementaries.common.block.tiles.SackBlockTile;
 import net.mehvahdjukaar.supplementaries.reg.ModTextures;
 
-public final class SackPreviewProvider extends BlockEntityPreviewProvider {
+public final class LunchBasketPreviewProvider extends BlockEntityPreviewProvider {
 
     @Environment(EnvType.CLIENT)
     private static PreviewRenderer renderer;
 
-    public SackPreviewProvider() {
-        super(SackBlockTile.getUnlockedSlots(), true);
+    public LunchBasketPreviewProvider() {
+        super(LunchBoxBlockTile.getUnlockedSlots(), true);
     }
 
     @Override
     public int getInventoryMaxSize(PreviewContext context) {
-        return SackBlockTile.getUnlockedSlots();
+        return LunchBoxBlockTile.getUnlockedSlots();
     }
 
     @Override
     @Environment(EnvType.CLIENT)
     public PreviewRenderer getRenderer() {
         if (renderer == null) {
-            renderer = new VariableSizePreviewRenderer(SackBlockTile::getUnlockedSlots, ModTextures.SACK_GUI_TEXTURE);
+            renderer = new VariableSizePreviewRenderer(LunchBoxBlockTile::getUnlockedSlots, ModTextures.LUNCH_BASKET_GUI_TEXTURE);
         }
         return renderer;
     }